requirements.txt
kivy>=2.1.0
kivymd>=1.1.1
supabase>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
